requirements.txt
faker
xlrd==1.2.0
PyMySQL==0.9.3
pytest==5.2.4
allure_python_commons==2.8.6
selenium==3.141.0
Appium_Python_Client==0.48
